user system set

+++ app/Http/Controllers/TasksController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Task;
+use Illuminate\Support\Facades\Auth;
 
 class TasksController extends Controller
 {
     public function index()
     {
-        $tasks = Task::all();
+        $user_id = Auth::id();
+        $tasks = Task::where('user_id', $user_id)->get();
         return view('tasks.index', [
             'tasks' => $tasks,    
         ]);
@@ -29,9 +31,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required',
         ]);
+        $user_id = Auth::id();
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = $user_id;
         $task->save();
         return redirect('/');
     }
@@ -39,7 +43,11 @@ class TasksController extends Controller
 
     public function show($id)
     {
+        $user_id = Auth::id();
         $task = Task::findOrFail($id);
+        if($user_id != $task->user_id){
+            return redirect('/');
+        }
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -47,7 +55,11 @@ class TasksController extends Controller
 
     public function edit($id)
     {
+        $user_id = Auth::id();
         $task = Task::findOrFail($id);
+        if($user_id != $task->user_id){
+            return redirect('/');
+        }
         return view('tasks.edit', [
             'task' => $task,
         ]);
@@ -59,7 +71,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required',
         ]);
+        $user_id = Auth::id();
         $task = Task::findOrFail($id);
+        if($user_id != $task->user_id){
+            return redirect('/');
+        }
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -68,7 +84,11 @@ class TasksController extends Controller
 
     public function destroy($id)
     {
+        $user_id = Auth::id();
         $task = Task::findOrFail($id);
+        if($user_id != $task->user_id){
+            return redirect('/');
+        }
         $task->delete();
         return redirect('/');
     }
